+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import random_toolkit
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  RandomToolkitPlugin.register(with: registry.registrar(forPlugin: "RandomToolkitPlugin"))
 }