+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,9 +6,6 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <random_toolkit/random_toolkit_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  RandomToolkitPluginCApiRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("RandomToolkitPluginCApi"));
 }